+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.20.3.2 - 3rd June 2020
 
+* more-complete tests for stock severity levels
 * minor documentation improvements
 
 ## previous versions